# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ERP Van Sales API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /storage/uploads/request-url:
    post:
      tags: [Storage]
      operationId: requestUploadUrl
      summary: Request a presigned URL for file upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadUrlRequest'
      responses:
        '200':
          description: Presigned upload URL generated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadUrlResponse'
        '400':
          description: Missing or invalid fields.
...
  /storage/objects/{objectPath}:
    get:
      tags: [Storage]
      operationId: getStorageObject
      summary: Serve an object entity
      parameters:
        - name: objectPath
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Object content streamed.
          content:
            '*/*':
              schema:
                type: string
                format: binary
        '404':
          description: Not found.

  # PRODUCTS
...
  /reports/daily:
    get:
      operationId: getDailyReport
      tags: [reports]
      summary: Get daily sales report
      parameters:
        - name: date
          in: query
          schema:
            type: string
            format: date
        - name: truckId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Daily report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyReport"
  /reports/trucks:
    get:
      operationId: getTruckPerformance
      tags: [reports]
      summary: Get truck performance report
      parameters:
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date
        - name: dateTo
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Truck performance
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TruckPerformance"
...
  /reports/top-products:
    get:
      operationId: getTopProducts
      tags: [reports]
      summary: Get top selling products
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Top products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TopProduct"
components:
  schemas:
...
    DailyReport:
      type: object
      properties:
        date:
          type: string
        totalSales:
          type: number
        cashSales:
          type: number
        creditSales:
          type: number
        totalCommission:
          type: number
        invoiceCount:
          type: integer
        topProducts:
          type: array
          items:
            $ref: "#/components/schemas/TopProduct"
        byTruck:
          type: array
          items:
            $ref: "#/components/schemas/TruckPerformance"
      required:
        - date
        - totalSales
        - cashSales
        - creditSales
        - totalCommission
        - invoiceCount
        - topProducts
        - byTruck
    TruckPerformance:
      type: object
      properties:
        truckId:
          type: integer
        truckName:
          type: string
        vendeurName:
          type: string
          nullable: true
        totalSales:
          type: number
        cashSales:
          type: number
        creditSales:
          type: number
        totalCommission:
          type: number
        invoiceCount:
          type: integer
      required:
        - truckId
        - truckName
        - totalSales
        - cashSales
        - creditSales
        - totalCommission
        - invoiceCount
...
    TopProduct:
      type: object
      properties:
        productId:
          type: integer
        productName:
          type: string
        totalQuantity:
          type: number
        totalRevenue:
          type: number
        invoiceCount:
          type: integer
      required:
        - productId
        - productName
        - totalQuantity
        - totalRevenue
        - invoiceCount
...
    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
          minLength: 1
        size:
          type: integer
          minimum: 1
        contentType:
          type: string
          minLength: 1
    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
          format: uri
        objectPath:
          type: string
        metadata:
          $ref: '#/components/schemas/UploadUrlRequest'